requirements.txt
fastapi==0.125.0
uvicorn[standard]==0.38.0
pymongo==4.15.5
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-dotenv==1.2.1
pydantic==2.12.5
pydantic-settings==2.6.1
email-validator==2.2.0

